backend/user_service/requirements.txt
fastapi
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
motor
uvicorn